requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
cssselect==1.0.3
ddict==0.1.3
Django==2.1.3
django-webpack-loader==0.6.0
djangorestframework==3.9.0
fake-useragent==0.1.11
idna==2.7
lxml==4.2.5
parse==1.9.0
pyquery==1.4.0
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
w3lib==1.19.0
xmltodict==0.11.0
